refactor eq poly
