+++ arith/mpoly/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "mpoly"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+arith = { path = "../" }
+
+ark-std.workspace = true
+halo2curves.workspace = true
