+++ sumcheck/Cargo.toml
@@ -10,5 +10,5 @@ circuit = { path = "../circuit" }
 config = { path = "../config" }
 transcript = { path = "../transcript" }
 
-env_logger = "0.11.3"
-log = "0.4"
+env_logger.workspace = true
+log.workspace = true
