+++ sumcheck/src/lib.rs
@@ -13,6 +13,3 @@ pub use scratch_pad::GkrScratchpad;
 
 // mod sumcheck_arbitrary;
 // pub use sumcheck_arbitrary::*;
-
-mod eq_poly;
-pub use eq_poly::*;
